Notification test button works
